+++ bin/funkoff.dart
@@ -1,0 +1,5 @@
+import 'package:funkoff/shellAPI.dart';
+
+void main(){
+  functionsShell();
+}