+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 - initial release to GitHub
 - initial upload to Pub.dev
+
+## Version 1.1.0: Functions shell
+
+- added an API for evaluating functions and datasets in a shell
+- added a shell binary